FilmCardPresenter is done
