+++ src/const.js
@@ -1,4 +1,4 @@
-export const FILMS_QUANTITY = 25;
+export const FILMS_QUANTITY = 7;
 export const FILMS_STEP = 5;
 
 export const SORTS = ['Sort by default', 'Sort by date', 'Sort by rating'];
